Move getMediaItems to a hook and optimise to only get the media for items we dont have yet
